# api_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GrantFlow AI API",
    "version": "1.0.0"
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Root",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/analyze": {
      "post": {
        "summary": "Analyze Bando",
        "operationId": "analyze_bando_analyze_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_bando_analyze_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/verify": {
      "post": {
        "summary": "Verify Eligibility",
        "operationId": "verify_eligibility_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/history": {
      "get": {
        "summary": "Get History",
        "operationId": "get_history_history_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/history/{analysis_id}": {
      "delete": {
        "summary": "Delete History",
        "operationId": "delete_history_history__analysis_id__delete",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_analyze_bando_analyze_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "visura": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Visura"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_analyze_bando_analyze_post"
      },
      "CompanyData": {
        "properties": {
          "ragione_sociale": {
            "type": "string",
            "title": "Ragione Sociale"
          },
          "ateco": {
            "type": "string",
            "title": "Ateco"
          },
          "dimensione": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dimensione",
            "default": ""
          },
          "regione": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regione",
            "default": ""
          },
          "fatturato": {
            "type": "number",
            "title": "Fatturato",
            "default": 0
          },
          "dipendenti": {
            "type": "integer",
            "title": "Dipendenti",
            "default": 0
          },
          "data_costituzione": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Costituzione",
            "default": ""
          },
          "investimento": {
            "type": "number",
            "title": "Investimento",
            "default": 0
          },
          "finanziamento_richiesto": {
            "type": "number",
            "title": "Finanziamento Richiesto",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "ragione_sociale",
          "ateco"
        ],
        "title": "CompanyData"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifyRequest": {
        "properties": {
          "dati_azienda": {
            "$ref": "#/components/schemas/CompanyData"
          },
          "parametri_finanziari": {
            "additionalProperties": true,
            "type": "object",
            "title": "Parametri Finanziari"
          },
          "scheda_bando": {
            "type": "string",
            "title": "Scheda Bando",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "dati_azienda",
          "parametri_finanziari"
        ],
        "title": "VerifyRequest"
      }
    }
  }
}